fixed the sort, the compare function returned an int instead of a float

+++ bonus/includes_bonus/cub3d_bonus.h
@@ -6,7 +6,7 @@
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 10:41:27 by mhotting          #+#    #+#             */
-/*   Updated: 2024/06/29 15:26:37 by brappo           ###   ########.fr       */
+/*   Updated: 2024/06/29 21:57:00 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -136,6 +136,6 @@ void		remove_last_spaces(char *str);
 void		skip_next_spaces(char **str);
 void		display_delta_time(void);
 long		get_tick(void);
-void		sort_list(t_list *lst, int compare(void *, void *));
+void		sort_list(t_list *lst, float compare(void *, void *));
 
 #endif
